pyproject.toml
[project]
name = "the-working-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.79.0",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
]
